pyproject.toml
[tool.poetry]
name = "fastapi-school"
version = "0.1.0"
description = ""
authors = ["Lev <lafitel31@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.112.2"
pydantic = {extras = ["email"], version = "^2.8.2"}
pydantic-settings = "^2.4.0"
sqlalchemy = {extras = ["asyncio"], version = "^2.0.32"}
uvicorn = "^0.30.6"
alembic = "^1.13.2"
asyncpg = "^0.29.0"
gunicorn = "^23.0.0"
pyjwt = {extras = ["crypto"], version = "^2.9.0"}
bcrypt = "^4.2.0"
python-multipart = "^0.0.9"


[tool.poetry.group.dev.dependencies]
black = "^24.8.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
